Limit the chunk size used for the dual numbers in ForwardDiff AD
Dual numbers with excessive chunk size are expensive to compile and they tend to be slow
at runtime due to the tuples becoming so large.

Also, functions have to be compiled from scratch when a new chunk size is encountered,
so if a program uses a large set of chunk sizes the compilation time can be significant.

There were examples in some models where the chunk size was over 30 which is way too big.
ForwardDiff itself limits the chunk size to 12 unless it is given explicitly.

Here, we limit the chunk size to 4 to reduce compilation time. It is likely that this has a negligble
impact on the full runtime but some benchmarking is warranted.

This takes the precompilation time for the FRB_US model from 32 s to 22s