Precompute looking up indices for all variables in `update_auxvars`
Looking up the index for a single variable is quite expensive since
we need to find that variable in a vector. The current method ended up
calling `indexin` which is not optimized well for finding a single value
since it does quite a bit of expensive stuff under the hood (like creating
a dictionary).

Instead, in this PR, we precompute a variable => index mapping for all variables
which is done via a simple loop and then use that mapping when an index is needed.

This takes the time for solving the FRB-US model (following the tutorial in TutorialsEcon and using
the `@btime` macro from BenchmarkTools.jl) from

```julia
julia> @btime simulate(m, p_0, ed_0; verbose=true, tol=1e-12);
  472.368 ms (1646933 allocations: 3.80 GiB)
```

to

```julia
julia> @btime simulate(m, p_0, ed_0; verbose=true, tol=1e-12);
  148.714 ms (1389881 allocations: 946.72 MiB)
```