add a `load_model` function and use it to allow loading multiple versions of models simultaneously
This is a proof of concept for a slightly more structured model loading than the
current one which is based on modifying the `LOAD_PATH` and loading the module
as a package. The issue with the current system is that you can not load
multiple models with the same name since the Julia namespace for files loaded as
packages is just the name of the module. However, using the Julia package system
for the models means that it can be precompiled which can generate big time
savings over multiple sessions where the same model is loaded.

This PR creates a new `load_model` function with the intention of keeping the
advantages of precompiling the models which also allow loading multiple models
of the same name. This is done by having an optional argument to `load_model`
that can give a separate "alias" for the model. When this is given, the `module
Model` part in the model is replaced with the alias. The model file is copied
to a scratch space before getting loaded in order to not pollute the folder where
they are loaded from.

The drawbacks to this approach are:
- Splitting a model into multiple files with `include` will not work out of the box.
- Stacktraces etc will also point to this generated file in the scratch space

Below is a usage example:

```julia
[ Info: Precompiling simple_RBC [top-level]
6 variable(s): C, K, L, w, r, A
1 shock(s): ea
7 parameter(s): g = 0.015, α = 0.33, γ = 0
    δ = 0.1, λ = 0.97, ρ = 0.03, β = @link 1 / (1 + ρ)
6 equations(s):
   E1:   @log 1 / C[t] = β * (1 / (C[t + 1] * (1 + g))) * ((r[t + 1] + 1) - δ)
   ...

In [3]: RBC2 = load_model("../2.simple_RBC/simple_RBC.jl", "simple_RBC2")
[ Info: Precompiling simple_RBC2 [top-level]
6 variable(s): C, K, L, w, r, A
1 shock(s): ea
7 parameter(s): g = 0.015, α = 0.33, γ = 0
    δ = 0.1, λ = 0.97, ρ = 0.03, β = @link 1 / (1 + ρ)
6 equations(s):
   E1:   @log 1 / C[t] = β * (1 / (C[t + 1] * (1 + g))) * ((r[t + 1] + 1) - δ)
...
```

Since this is the first time the model is loaded it has to be precompiled.
And a model with a different alias also need to be precompiled.

In a follow up session, the models are already precompiled and load quickly:

```julia
julia> RBC = load_model("../2.simple_RBC/simple_RBC.jl");

julia> RBC2 = load_model("../2.simple_RBC/simple_RBC.jl", "simple_RBC2");
```