tighten up types in the `params` for `EquationEvaluator`
The current parameters have a declared type of `Any` in the struct which leads to bad performance.
For example, loading the FRB-US model and looking at code_warntype

This PR tightens up the allowed types in the `EquationEvaluator` for the `params` to allow stronger inference and union splitting.

Running the follwoing benchmark:

```
unique!(push!(LOAD_PATH, realpath("./models")))
using ModelBaseEcon
using FRBUS_VAR

m = FRBUS_VAR.model
nrows = 1 + m.maxlag + m.maxlead
ncols = length(m.allvars)
pt = zeros(nrows, ncols);
using BenchmarkTools
@btime @eval eval_RJ(pt, m);
```

The time goes from

`544.104 μs (22094 allocations: 860.48 KiB)`
to
`326.093 μs (5581 allocations: 228.33 KiB)`

The drawback of course is that only these type of parameters can be used in a model.
If a restriction to the parameters is ok, something like this can be done.
If the parameters truly need to be able to take on any type, we need to rethink a little bit in using a single
callable struct for all equations.